requirement.txt
openai
langchain
streamlit
python-dotenv
langchain_community
langchain_openai